User form action refactored

+++ app/view/user/index.php
@@ -5,7 +5,7 @@
 </div>
 <?php } ?>
 <h4>Lista użytkowników</h4>
-<a class="add" href="?cmd=user-form">Nowy użytkownik</a>
+<a class="add" href="?cmd=user-new">Nowy użytkownik</a>
 <table class="list_table">
     <tr>
         <th>Login</th>
@@ -15,17 +15,21 @@
     </tr>
     <?php foreach ($users as $user) {?>
     <tr>
-        <td class="feature"><?php echo $user->getNick();?></td>
+        <td class="feature"><?php echo $user->getUsername();?></td>
         <td class="feature"><?php echo $user->getFirstName();?></td>
         <td class="feature"><?php echo $user->getLastName();?></td>
         <td class="feature"><?php echo $user->getEmail();?></td>
         <td>
-            <a class="delete" href="?cmd=user-delete&id=<?php echo $user->getId();?>">
-                Usuń
-            </a>
-            <a class="edit" href="?cmd=user-form&id=<?php echo $user->getId();?>">
+            <?php if (isset($buttons['user-edit'])) {?>
+            <a class="edit" href="?cmd=user-edit&id=<?php echo $user->getId();?>">
                 Edytuj
             </a>
+            <?php } ?>
+            <?php if (isset($buttons['user-delete'])) {?>
+                <a class="delete" href="?cmd=user-delete&id=<?php echo $user->getId();?>">
+                    Usuń
+                </a>
+            <?php } ?>
         </td>
     </tr>
     <?php } ?>
